chore(sanity): rename and change structure for components related to document head version
